clean up, defer model loading, cache model and db
